+++ packages/client/src/play/index.js
@@ -1,3 +1,6 @@
 export { createLivePlayClient } from './live.js'
 export * from './chrome.js'
+export * from './occupancy.js'
+export * from './sidebar-model.js'
+export { PlayWorkspaceBrowser } from './sidebar.js'
 export * from './schema.js'